+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.0.2 (unreleased)
+1.0.2 (2022-04-25)
 ------------------
 
 - Hide unwanted upgrades from site-creation and quickinstaller